requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.9
fonttools==4.56.0
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
tzdata==2025.1
uvicorn==0.34.0
wheel==0.45.1
